Move Mesh from DataStructures to Domain
