+++ src/NumericalAlgorithms/LinearOperators/CMakeLists.txt
@@ -15,6 +15,7 @@ add_library(${LIBRARY} ${LIBRARY_SOURCES})
 target_link_libraries(
   ${LIBRARY}
   INTERFACE DataStructures
+  INTERFACE Domain
   INTERFACE ErrorHandling
   INTERFACE Spectral
   )
