+++ src/NumericalAlgorithms/Spectral/CMakeLists.txt
@@ -13,5 +13,6 @@ add_library(${LIBRARY} ${LIBRARY_SOURCES})
 target_link_libraries(
   ${LIBRARY}
   INTERFACE DataStructures
+  INTERFACE Domain
   INTERFACE ErrorHandling
   )
