+++ tests/Unit/DataStructures/CMakeLists.txt
@@ -16,7 +16,6 @@ set(LIBRARY_SOURCES
   Test_Identity.cpp
   Test_Index.cpp
   Test_IndexIterator.cpp
-  Test_Mesh.cpp
   Test_OrientVariablesOnSlice.cpp
   Test_SliceIterator.cpp
   Test_StripeIterator.cpp
