+++ tests/Unit/Domain/CMakeLists.txt
@@ -22,6 +22,7 @@ set(LIBRARY_SOURCES
   Test_InitialElementIds.cpp
   Test_InterfaceItems.cpp
   Test_LogicalCoordinates.cpp
+  Test_Mesh.cpp
   Test_Neighbors.cpp
   Test_OrientationMap.cpp
   Test_SegmentId.cpp
